[Finishes #166702465] Refactored codes
