+++ server/api/utils/imageUploader.js
@@ -1,6 +1,12 @@
 import faker from 'faker';
 import uploader from '../config/cloudinary';
 
+/**
+ * @function imageUploader
+ * @description - uploads images to cloudinary
+ * @param {object} image - image to be uploaded
+ * @exports imageUploader
+ */
 const imageUploader = async (image) => {
   let imageUrl = '';
   if (process.env.NODE_ENV !== 'production') {
